Problem Number 6,9,11
### content_suppliers_list
6# When the user is searching for something, each change in the input
field should be a search. So, the result will be filtered until the user
finds what he wants or close the modal.

### content_listings_list
9# When the user is searching for something, each change in the input
field should be a search. So, the result will be filtered until the user
finds what he wants or close the modal.

### content_products_select
- 11# When the user is searching for something, each change in the input
field should be a search. So, the result will be filtered until the user
finds what he wants or close the modal.